Subclass leadboard functionality for users
