+++ app/controllers/community_controller.rb
@@ -1,9 +1,9 @@
 class CommunityController < ApplicationController
-  actions_without_auth :leaderboards
+  actions_without_auth :user_leaderboards
 
-  def leaderboards
-    data = Rails.cache.fetch('community_leaderboards', expires_in: 10.minutes) do
-      leaderboard = Leaderboard.new(UserPresenter)
+  def user_leaderboards
+    data = Rails.cache.fetch('community_user_leaderboards', expires_in: 10.minutes) do
+      leaderboard = Leaderboard::UserLeaderboard.new(UserPresenter)
       leaderboard.get
     end
     render json: data
